% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{attribute_tibble_to_list}
\alias{attribute_tibble_to_list}
\title{Helper function to convert attributes to list}
\usage{
attribute_tibble_to_list(tb)
}
\arguments{
\item{tb}{tibble/data.frame. First column is the original value, the second the new label}
}
\value{
list object
}
\description{
Little helper function that converts tibble into a list with vectors, which is the
expected attributes input for get_census_summary()
}
