% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auscensus-package.R
\docType{package}
\name{auscensus-package}
\alias{auscensus}
\alias{auscensus-package}
\title{auscensus: Access Australian Census Data (2006-2021)}
\description{
R package to interact with Australian Census Data Packs,providing an interface to extract data across multiple censuses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://carlosyanez.github.io/auscensus/}
  \item Report bugs at \url{https://github.com/carlosyanez/auscensus/issues}
}

}
\author{
\strong{Maintainer}: Carlos Y치침ez Santib치침ez \email{carlos.yanez.s@gmail.com}

Other contributors:
\itemize{
  \item Craig Alexander \email{Craig.Alexander.2@glasgow.ac.uk} [thesis advisor]
  \item Kyle Walker \email{kyle@walker-data.com} [copyright holder]
  \item Australian Bureau of Statistics [copyright holder]
}

}
\keyword{internal}
