% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_percentage}
\alias{calculate_percentage}
\title{Helper function to convert attributes to list}
\usage{
calculate_percentage(
  df,
  key_col,
  value_col,
  key_value = "Total",
  percentage_scale = 1
)
}
\arguments{
\item{df}{data frame}

\item{key_col}{name of the column containing the "Total Label"}

\item{value_col}{name of the column containing values}

\item{key_value}{total label}

\item{percentage_scale}{1 if percentage to be presented in scale 0-1, or 100 to be shown as 0\%-100\%}
}
\value{
list object
}
\description{
Little helper function that converts tibble into a list with vectors, which is the
expected attributes input for get_census_summary()
}
\keyword{helpers}
