% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{census_datapacks}
\alias{census_datapacks}
\title{Helper function to download  data}
\usage{
census_datapacks()
}
\value{
nothing
}
\description{
Helper function to download  data
}
\keyword{helpers}
