% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{data_census_delete}
\alias{data_census_delete}
\title{Helper function to update/download  data}
\usage{
data_census_delete(file = NULL)
}
\arguments{
\item{file}{to delete - defaults to all of them (provide full path, can obtain from data_census_info)}
}
\value{
nothing
}
\description{
Helper function to update/download  data
}
\keyword{helpers}
