% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{data_census_download}
\alias{data_census_download}
\title{Helper function to download  data}
\usage{
data_census_download(
  download_dir,
  census_year = NULL,
  download_method = "wget"
)
}
\arguments{
\item{download_dir}{Full path where to download census files (required)}

\item{census_year}{census year to download (default to all)}

\item{download_method}{method to pass to download.file() ("wget" as default)}
}
\value{
nothing
}
\description{
Helper function to download  data
}
\keyword{helpers}
