% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{data_census_info}
\alias{data_census_info}
\title{Helper function to update/download  data}
\usage{
data_census_info()
}
\value{
nothing
}
\description{
Helper function to update/download  data
}
\keyword{helpers}
