% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_census_summary}
\alias{get_census_summary}
\title{Obtain get data from filename}
\usage{
get_census_summary(
  table_number,
  geo_structure,
  attributes,
  geo_units = NULL,
  selected_years = list_census_years(),
  reference_total = NULL,
  cache = TRUE,
  ignore_cache = FALSE
)
}
\arguments{
\item{table_number}{number of selected table}

\item{geo_structure}{vector with strings of geo structures (e.g. SA1,LGA,CED)}

\item{attributes}{attribure}

\item{geo_units}{geo units}

\item{selected_years}{years to filter}

\item{reference_total}{reference total}

\item{cache}{if TRUE, will save the query in the cache (in parquet file) for later use}

\item{ignore_cache}{If TRUE, it will ignore cached files}
}
\value{
data frame with data from file, filtered by division and election year
}
\description{
Obtain get data from filename
}
\keyword{getdata}
