% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_census_attributes}
\alias{list_census_attributes}
\title{Get names of attributes for a given census tables, across all time}
\usage{
list_census_attributes(number = NULL, attribute_regex = NULL)
}
\arguments{
\item{number}{vector containing one or more table numbers}

\item{attribute_regex}{string with a regular expression to filter attribute names}
}
\value{
tibble, showing the geo type, available for each year
}
\description{
Get list of available geopgrahies, filterable by type and name.
}
\examples{
\dontrun{
# Get list of all divisions
list_census_attributes()
 }
}
\keyword{lists}
