% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_census_geo}
\alias{list_census_geo}
\title{Get census geographies, filterable}
\usage{
list_census_geo(geo_types = NULL, geo_names = NULL, geo_name_regex = NULL)
}
\arguments{
\item{geo_types}{vector containing one or more geography types (i.e. "STE","CED","SA1" ). NULL by default.}

\item{geo_names}{vector containing one or more geography names (i.e. "Melbourne", "Yarra","Stonnington" for LGAs). NULL by default.}

\item{geo_name_regex}{string with a regular expression to filter geograhpy names (i.e. for all elements starting with M : "$M")}
}
\value{
tibble, showing the geo type, available for each year
}
\description{
Get list of available geopgrahies, filterable by type and name.
}
\examples{
\dontrun{
# Get list of all divisions
list_census_geo()
 }
}
\keyword{lists}
