% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_census_geo_tables}
\alias{list_census_geo_tables}
\title{List if a geo structure is available for a particular table , in a particular year}
\usage{
list_census_geo_tables(year = NULL, geo = NULL, table_number = NULL)
}
\arguments{
\item{year}{vector with years}

\item{geo}{vector with geo structure}

\item{table_number}{table number}
}
\value{
tibble
}
\description{
List if a geo structure is available for a particular table , in a particular year
}
\examples{
\dontrun{
# Get list of all divisions
list_census_geo()
 }
}
\keyword{lists}
