% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_census_geo_types}
\alias{list_census_geo_types}
\title{Get Geography types.}
\usage{
list_census_geo_types()
}
\value{
tibble, showing the geotype, available for each year
}
\description{
Very simple function listing geography types (e.g. SAx, CED, etc.), for which data pack has been imported.
}
\examples{
\dontrun{
# Get list of all divisions
list_census_geo_types()
 }
}
\keyword{lists}
