% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_census_tables}
\alias{list_census_tables}
\title{Get census geographies, filterable}
\usage{
list_census_tables(number = NULL, table_name_regex = NULL)
}
\arguments{
\item{number}{vector containing one or more table numbers}

\item{table_name_regex}{string with a regular expression to filter table names (i.e. for all elements containing with Country : "Country")}
}
\value{
tibble, showing the geo type, available for each year
}
\description{
Get list of available geopgrahies, filterable by type and name.
}
\examples{
\dontrun{
# Get list of all divisions
list_census_geo()
 }
}
\keyword{lists}
