% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_census_years}
\alias{list_census_years}
\title{Get Census years.}
\usage{
list_census_years(mode = "available")
}
\arguments{
\item{mode}{Either "listed" or "available}
}
\value{
vector with years
}
\description{
Very simple function listing the Census years included in this package, for which data pack has been imported.
}
\examples{
\dontrun{
# Get list of all divisions
list_census_years()
 }
}
\keyword{lists}
