% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{remove_census_cache_csv}
\alias{remove_census_cache_csv}
\title{Helper function to delete all csv in cache}
\usage{
remove_census_cache_csv()
}
\value{
nothing
}
\description{
Helper function to delete all csv in cache
}
\keyword{helpers}
