% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_census_data}
\alias{get_census_data}
\title{Obtain get data from filename}
\usage{
get_census_data(
  census_table,
  geo_structure,
  selected_years = list_census_years(),
  cache = TRUE,
  ignore_cache = FALSE
)
}
\arguments{
\item{census_table}{list of tables, in the format of the output of list_census_tables()}

\item{geo_structure}{vector with strings of geo structures (e.g. SA1,LGA,CED)}

\item{selected_years}{years to filter}

\item{cache}{if TRUE, will save the query in the cache (in parquet file) for later use}

\item{ignore_cache}{If TRUE, it will ignore cached files}
}
\value{
data frame with data from file, filtered by division and election year
}
\description{
Obtain get data from filename
}
\keyword{getdata}
