% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_census_data}
\alias{get_census_data}
\title{Get census data.}
\usage{
get_census_data(
  census_table,
  geo_structure,
  selected_years = list_census_years(),
  ignore_cache = FALSE,
  collect_data = FALSE,
  attr = NULL
)
}
\arguments{
\item{census_table}{list of tables, in the format of the output of list_census_tables()}

\item{geo_structure}{vector with strings of geo structures (e.g. SA1,LGA,CED)}

\item{selected_years}{years to filter}

\item{ignore_cache}{If TRUE, it will ignore cached files}

\item{collect_data}{if TRUE  will return data. if FALSE (default) , it will return {arrow} bindings to cached files}

\item{attr}{attributes to filter on, presented as a character vector (e.g c("Age_years_60_males","Age_years_60_females"))}
}
\value{
data frame with data from file, filtered by division and election year
}
\description{
This function extracts table files from each data pack (given tables and geo structure), and will collate them together into a list(),
which it will return. By default it will save the processed tables in the cache folder (in parquet files), which it will use on subsquent calls.
}
\examples{
\dontrun{
data <- get_census_data(census_table = list_census_tables("04"),
                        geo_structure = "LGA")

names(data)
}
}
\keyword{getdata}
