% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_census_summary}
\alias{get_census_summary}
\title{Get a summary of one or a collection of statistics across Censuses}
\usage{
get_census_summary(
  table_number = NULL,
  geo_structure = NULL,
  attribute,
  geo_unit_names = NULL,
  geo_unit_codes = NULL,
  selected_years = list_census_years(),
  reference_total = NULL,
  percentage_scale = 1,
  ignore_cache = FALSE,
  data_source = NULL,
  data_collected = FALSE,
  census_table = NULL
)
}
\arguments{
\item{table_number}{number of selected table}

\item{geo_structure}{character presenting the geographical structure to present stats (e.g. SA1,LGA,CED)}

\item{attribute}{list with vectors of statistics to be summarise. Each vector element will be aggregated and presented under the item's name, e.g.
list("60 year old male"=c("Age_years_60_males","Age (Years): 60_males"),}

\item{geo_unit_names}{vector with names of the geographic structures to present. They need to correspond with geo_structure, e.g. if geo_structure="LGA",
acceptable values could be  c("Melbourne","Stonnington","Yarra"). If both this and geo_unit_codes are null, it will present all avaialable elements.}

\item{geo_unit_codes}{vector with ABS codes of the geo structures to present. Similar to geo_units_names.}

\item{selected_years}{vector with selected years to display.}

\item{reference_total}{Optional. List containing the names of all statistics representing totals, e.g. list("Total"=c("Total_persons")}

\item{percentage_scale}{1 if percentage to be presented in scale 0-1, or 100 to be shown as 0\%-100\%}

\item{ignore_cache}{If TRUE, it will ignore cached files}

\item{data_source}{result of get_census_data (will ignore other parameters if this is provided)}

\item{data_collected}{TRUE if data_source is a dataset, FALSE if is  a DB,arrow binding}

\item{census_table}{Instead of using a table number, this allows for a more complex filter table, e.g. containing different table numbers.
Expected format matches the output of list_census_table().}
}
\value{
data frame with data from file, filtered by division and election year
}
\description{
This function allows to produce a summary of one or many statistics across censuses. Results are presented in a simple summary table.
The function allows to present individual statistics or an aggregation of several statistics (e.g. aggregate number of births by country to present
a continental total). If the name statistic containing totals is provided, the function has an option to calculate percentages (presented either in 0-1 or 0-100 scale).
}
\examples{
\dontrun{
get_census_summary(table_number = "04",
    attribute = list("60 year old male"=c("Age_years_60_males","Age (Years): 60_males"),
                      "60 year old female"=c("Age_years_60_males","Age (Years): 60_females")),
    geo_unit_names = c("Melbourne","Stonnington","Yarra"),
    reference_total = list("Total"=c("Total_persons")),

}
}
\keyword{getdata}
