% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_census_summary}
\alias{get_census_summary}
\title{Get a particular data point across census}
\usage{
get_census_summary(
  table_number = NULL,
  geo_structure = NULL,
  attribute,
  geo_unit_names = NULL,
  geo_unit_codes = NULL,
  selected_years = list_census_years(),
  reference_total = NULL,
  percentage_scale = 1,
  ignore_cache = FALSE,
  data_source = NULL,
  data_collected = FALSE,
  census_table = NULL
)
}
\arguments{
\item{table_number}{number of selected table}

\item{geo_structure}{vector with strings of geo structures (e.g. SA1,LGA,CED)}

\item{attribute}{attribute}

\item{geo_unit_names}{geo unit names}

\item{geo_unit_codes}{geo unit codes}

\item{selected_years}{years to filter}

\item{reference_total}{reference total}

\item{percentage_scale}{1 if percentage to be presented in scale 0-1, or 100 to be shown as 0\%-100\%}

\item{ignore_cache}{If TRUE, it will ignore cached files}

\item{data_source}{result of get_census_data (will ignore other parameters if this is provided)}

\item{data_collected}{TRUE if data_source is a dataset, FALSE if is  a DB,arrow binding}

\item{census_table}{output of list_census_table()}
}
\value{
data frame with data from file, filtered by division and election year
}
\description{
Extracts a particular data point statistics from census data packs, filterable by particular geographical units.
Provides and option to express values as percentage of another data point.
}
\examples{
\dontrun{
get_census_summary(table_number = "04",
    attribute = list("60 year old male"=c("Age_years_60_males","Age (Years): 60_males"),
                      "60 year old female"=c("Age_years_60_males","Age (Years): 60_females")),
    geo_unit_names = c("Melbourne","Stonnington","Yarra"),
    reference_total = list("Total"=c("Total_persons")),

}
}
\keyword{getdata}
