% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_census_summary}
\alias{get_census_summary}
\title{Get a particular data point across census}
\usage{
get_census_summary(
  table_number = NULL,
  geo_structure = NULL,
  attributes,
  geo_units = NULL,
  selected_years = list_census_years(),
  reference_total = NULL,
  percentage_scale = 1,
  cache = TRUE,
  ignore_cache = FALSE,
  data_source = NULL
)
}
\arguments{
\item{table_number}{number of selected table}

\item{geo_structure}{vector with strings of geo structures (e.g. SA1,LGA,CED)}

\item{attributes}{attribute}

\item{geo_units}{geo units}

\item{selected_years}{years to filter}

\item{reference_total}{reference total}

\item{percentage_scale}{1 if percentage to be presented in scale 0-1, or 100 to be shown as 0\%-100\%}

\item{cache}{if TRUE, will save the query in the cache (in parquet file) for later use}

\item{ignore_cache}{If TRUE, it will ignore cached files}

\item{data_source}{result of get_census_data (will ignore other parameters if this is provided)}
}
\value{
data frame with data from file, filtered by division and election year
}
\description{
Extracts a particular data point statistics from census data packs, filterable by particular geographical units.
Provides and option to express values as percentage of another data point.
}
\examples{
\dontrun{
get_census_summary(table_number = "04",
    attributes = list("60 year old male"=c("Age_years_60_males","Age (Years): 60_males"),
                      "60 year old female"=c("Age_years_60_males","Age (Years): 60_females")),
    geo_units = c("Melbourne","Stonnington","Yarra"),
    reference_total = list("Total"=c("Total_persons")),

}
}
\keyword{getdata}
